requirements.txt
antiorm==1.2.1
appdirs==1.4.4
astroid==2.4.2
awscli==1.18.127
backcall==0.2.0
beautifulsoup4==4.9.3
black==20.8b1
botocore==1.17.50
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextvars==2.4
coverage==5.3
cryptography==3.2.1
decorator==4.4.2
Django==2.0.2
dnspython==1.16.0
docutils==0.15.2
eventlet==0.28.0
Flask==1.1.2
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.1
git-remote-codecommit==1.15.1
google==3.0.0
google-oauth==1.0.1
googletrans==3.0.0
greenlet==0.4.17
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.10.6
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ikp3db==1.4.1
immutables==0.14
importlib-metadata==1.7.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.4.2
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
mypy-extensions==0.4.3
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
parso==0.1.1
pathspec==0.8.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
prettytable==0.7.2
prompt-toolkit==3.0.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pybars3==0.9.7
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.6.0
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.6
PyMeta3==0.5.1
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rfc3987==1.3.8
rsa==4.5
s3transfer==0.3.3
six==1.15.0
sniffio==1.2.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
stevedore==3.2.0
toml==0.10.1
traitlets==4.3.3
tweepy==3.9.0
twilio==6.45.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==16.2.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
